Added tier 2 automove
Slight issue: my grid check function looks for obstructions at the square BEFORE the car moves. This means that there are some pathfinding errors which cannot be resolved until this issue is fixed.

It should be relatively simple to fix.

Added a directional updator after automove occurs, to ensure the car doesn't end up in a loop of being stuck going back and forth.