Added in game UI, fuel and death conditions
Added fuel, draining each movement tick. When it reaches zero the player is killed. Represented by a yellow bar in the newly added UI.

The UI consists of a 300 pixel wide bar on the right side of the screen, similar to that of the original Rally-X.
It has a score counter (with an improved font), a fuel counter, and a fuel gauge as of present.

Widened graphics frame for UI.

Added explosion graphic and condition to display it instead of the player when they died.

Added logic to deactivate player after death.

Added enemy counter data to grid datastructure and added to mapdata.rtxm.